Create map MeshLibrary. Move PathLayer scripts to dedicated directory.
