+++ Game/Logic/path_layer.tscn
@@ -1,8 +1,10 @@
 [gd_scene format=4 uid="uid://dvcjxpgeb5iyt"]
 
-[ext_resource type="Script" uid="uid://dwtejndhja206" path="res://Game/Logic/CompositePathLayer.cs" id="1_qnxpf"]
 [ext_resource type="Texture2D" uid="uid://3fkqwl8gm8mf" path="res://Untitled.png" id="1_t0mgb"]
-[ext_resource type="Script" uid="uid://c3vki86g2rcug" path="res://Game/Logic/PathLayer2D.cs" id="2_qnxpf"]
+[ext_resource type="Script" uid="uid://cxd86kromaxdx" path="res://Game/PathLayer/CompositePathLayer.cs" id="1_y1wtc"]
+[ext_resource type="Script" uid="uid://dws6d172p6if3" path="res://Game/PathLayer/PathLayer2D.cs" id="3_h2hc4"]
+[ext_resource type="MeshLibrary" uid="uid://qkwhbw2blv7l" path="res://Game/MapMeshes.tres" id="4_ib8oh"]
+[ext_resource type="Script" uid="uid://cc2vvqbh4jgpi" path="res://Game/PathLayer/PathLayer3D.cs" id="5_qnksd"]
 
 [sub_resource type="TileSetAtlasSource" id="TileSetAtlasSource_ga5si"]
 texture = ExtResource("1_t0mgb")
@@ -16,13 +18,21 @@ texture_region_size = Vector2i(64, 64)
 tile_size = Vector2i(64, 64)
 sources/0 = SubResource("TileSetAtlasSource_ga5si")
 
-[node name="CompositePathLayer" type="Node" unique_id=1175408720]
-script = ExtResource("1_qnxpf")
-metadata/_custom_type_script = "uid://dwtejndhja206"
+[node name="CompositePathLayer" type="Node" unique_id=499868216]
+script = ExtResource("1_y1wtc")
+metadata/_custom_type_script = "uid://cxd86kromaxdx"
 
-[node name="PathLayer" type="TileMapLayer" parent="." unique_id=1226762570]
+[node name="Instance2D" type="TileMapLayer" parent="." unique_id=1226762570]
 tile_map_data = PackedByteArray("AAAAAAcAAAAAAAAAAAABAAcAAAAAAAAAAAACAAcAAAAAAAAAAAADAAcAAAAAAAAAAAAEAAcAAAAAAAAAAAAEAAYAAAAAAAAAAAAFAAYAAAAAAAAAAAAGAAYAAAAAAAAAAAAGAAUAAAAAAAAAAAAGAAMAAAAAAAAAAAAFAAMAAAAAAAAAAAAFAAIAAAAAAAAAAAAEAAIAAAAAAAAAAAADAAIAAAAAAAAAAAACAAIAAAAAAAAAAAABAAIAAAAAAAAAAAAAAAIAAAAAAAAAAAAAAAMAAAAAAAAAAAAGAAQAAAAAAAAAAAA=")
 tile_set = SubResource("TileSet_ga5si")
-script = ExtResource("2_qnxpf")
+script = ExtResource("3_h2hc4")
 
-[node name="GridMap" type="GridMap" parent="." unique_id=213080206]
+[node name="Instance3D" type="GridMap" parent="." unique_id=774875888]
+mesh_library = ExtResource("4_ib8oh")
+cell_size = Vector3(2, 0.4, 2)
+cell_center_y = false
+data = {
+"cells": PackedInt32Array(7, 65526, 655360, 6, 65526, 655360, 5, 65526, 655360, 4, 65526, 655360, 7, 65527, 655360, 6, 65527, 655360, 5, 65527, 655360, 4, 65527, 655360, 7, 65528, 655360, 6, 65528, 655360, 5, 65528, 655360, 4, 65528, 655360, 8, 65528, 655360, 8, 65527, 655360, 8, 65526, 655360, 65540, 65528, 655361, 65540, 65527, 655361, 65540, 65526, 655361, 65541, 65526, 655362, 65541, 65527, 655362, 65541, 65528, 655362, 65543, 65526, 655365, 65543, 65527, 655365)
+}
+script = ExtResource("5_qnksd")
+metadata/_editor_floor_ = Vector3(0, 0, 0)
